refactor(storage): Improve memory store architecture with adapter pattern
Refactor memory store implementations to use a consistent adapter pattern with improved ID handling and vector search capabilities.

- Create abstract BaseMemoryStore class with consistent ID handling
- Implement StandardMemoryStore, ChunkedMemoryStore, and HybridMemoryStore
- Add adapter classes for each store type
- Ensure consistent interface across all adapters
- Fix ID resolution issues that were causing retrieval errors
- Add dedicated vector search component with multiple implementations
- Create factory with Pydantic configuration
- Fixed all imports to use newly refactored class(es):

This refactoring better separates storage concerns from access patterns and provides
a more robust framework for different memory store types while addressing the ID
resolution problems in the original implementation.

Part of MemoryWeave architectural improvements to support contextual fabric memory approach.
