+++ run_benchmark.py
@@ -54,7 +54,7 @@ from memoryweave.evaluation.synthetic.benchmark import (
     SyntheticBenchmark,
 )
 from memoryweave.interfaces.retrieval import Query, QueryType
-from memoryweave.storage.memory_store import Memory, MemoryStore
+from memoryweave.interfaces.memory import Memory, MemoryStore
 
 # set up rich logging
 logging.basicConfig(
@@ -369,7 +369,8 @@ class UnifiedBenchmark:
             )
 
             # Initialize memory manager
-            memory_store = MemoryStore()
+            memory_store = StandardMemoryStore()
+memory_adapter = MemoryAdapter(memory_store)
             memory_store.add_multiple(memories)
             memory_manager = MemoryManager(memory_store=memory_store)
